removed http address override from config file reader
