+++ bootstrap/signer/main/main.go
@@ -33,11 +33,13 @@ func main() {
 		return
 	}
 
-	cfg, err := config.GetNodeConfigFromFiles(configFiles, *httpAddress)
+	cfg, err := config.NewFromMultipleFiles(configFiles)
 	if err != nil {
 		fmt.Printf("%s \n", err)
 		os.Exit(1)
 	}
+
+	cfg.SetString(config.HTTP_ADDRESS, *httpAddress)
 
 	signer.StartSignerServer(cfg, getLogger()).WaitUntilShutdown()
 }
